requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
astroid==2.4.2
boto3==1.16.40
botocore==1.19.40
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django==3.1.4
django-storages==1.10.1
django-summernote==0.8.11.6
django-xhtml2pdf==0.0.4
djangorestframework==3.12.2
future==0.18.2
html5lib==1.1
idna==2.10
install==1.3.4
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
reportlab==3.5.56
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
